Move ProblemMiddleware to Core (as ExceptionMiddleware), with improvements
